passes on 64 bit now
git-svn-id: http://code.sixapart.com/svn/memcached@393 b0b603af-a30f-0410-a34e-baf09ae79d0b

+++ branches/memcached-1.1.x/memcached.c
@@ -358,7 +358,7 @@ void process_stat(conn *c, char *command) {
         pos += sprintf(pos, "STAT get_misses %u\r\n", stats.get_misses);
         pos += sprintf(pos, "STAT bytes_read %llu\r\n", stats.bytes_read);
         pos += sprintf(pos, "STAT bytes_written %llu\r\n", stats.bytes_written);
-        pos += sprintf(pos, "STAT limit_maxbytes %u\r\n", settings.maxbytes);
+        pos += sprintf(pos, "STAT limit_maxbytes %llu\r\n", (unsigned long long) settings.maxbytes);
         pos += sprintf(pos, "END");
         out_string(c, temp);
         return;
@@ -1549,7 +1549,7 @@ int main (int argc, char **argv) {
             settings.port = atoi(optarg);
             break;
         case 'm':
-            settings.maxbytes = atoi(optarg)*1024*1024;
+            settings.maxbytes = ((size_t)atoi(optarg))*1024*1024;
             break;
         case 'M':
             settings.evict_to_free = 0;
